feat(file-structure) outsource routes to different files for separation of concerns and easier management
